+++ gitgud/skills/basics/__init__.py
@@ -1,9 +1,12 @@
-from gitgud.skills.util import BasicLevel
 from gitgud.skills.util import Skill
+from gitgud.skills.level_builder import BasicLevel
 
 skill = Skill(
     'basics',
     [
-        BasicLevel(level_name, __name__) for level_name in ['committing', 'branching', 'merging', 'rebasing']
+        BasicLevel('committing', __name__),
+        BasicLevel('branching', __name__),
+        BasicLevel('merging', __name__),
+        BasicLevel('rebasing', __name__)
     ]
 )
