+++ gitgud/skills/intro/_intro/goal.txt
@@ -1,0 +1,3 @@
+Welcome to Git Gud!
+Type 
+    $ git gud test
