+++ gitgud/skills/intro/_intro/setup.spec
@@ -1,0 +1,2 @@
+1 (master)
+master
