+++ gitgud/skills/rampup/__init__.py
@@ -1,9 +1,12 @@
-from gitgud.skills.util import BasicLevel
 from gitgud.skills.util import Skill
+from gitgud.skills.level_builder import BasicLevel
 
 skill = Skill(
     'rampup',
     [
-        BasicLevel(level_name, __name__) for level_name in ['detaching', 'relrefs1', 'relrefs2', 'reversing']
+        BasicLevel('detaching', __name__),
+        BasicLevel('relrefs1', __name__),
+        BasicLevel('relrefs2', __name__),
+        BasicLevel('reversing', __name__)
     ]
 )
